+++ src/Data/Morpheus/Server/TH/Compile.hs
@@ -4,7 +4,6 @@
 module Data.Morpheus.Server.TH.Compile
   ( compileDocument,
     gqlDocument,
-    gqlDocumentNamespace,
   )
 where
 
@@ -13,11 +12,11 @@ where
 
 import qualified Data.ByteString.Lazy.Char8 as LB
 import Data.ByteString.Lazy.Char8 (ByteString)
-import Data.Morpheus.Server.CodeGen.Transform
+import Data.Morpheus.CodeGen
   ( parseServerTypeDefinitions,
   )
-import Data.Morpheus.Server.CodeGen.Types
-  ( ServerDecContext (..),
+import Data.Morpheus.CodeGen.Internal.AST
+  ( CodeGenConfig (..),
   )
 import Data.Morpheus.Server.TH.Declare
   ( runDeclare,
@@ -26,13 +25,10 @@ import Language.Haskell.TH
 import Language.Haskell.TH.Quote (QuasiQuoter (..))
 import Relude hiding (ByteString)
 
-gqlDocumentNamespace :: QuasiQuoter
-gqlDocumentNamespace = mkQuasiQuoter ServerDecContext {namespace = True}
-
 gqlDocument :: QuasiQuoter
-gqlDocument = mkQuasiQuoter ServerDecContext {namespace = False}
+gqlDocument = mkQuasiQuoter CodeGenConfig {namespace = False}
 
-mkQuasiQuoter :: ServerDecContext -> QuasiQuoter
+mkQuasiQuoter :: CodeGenConfig -> QuasiQuoter
 mkQuasiQuoter ctx =
   QuasiQuoter
     { quoteExp = notHandled "Expressions",
@@ -44,5 +40,5 @@ mkQuasiQuoter ctx =
     notHandled things =
       error $ things <> " are not supported by the GraphQL QuasiQuoter"
 
-compileDocument :: ServerDecContext -> ByteString -> Q [Dec]
+compileDocument :: CodeGenConfig -> ByteString -> Q [Dec]
 compileDocument ctx = parseServerTypeDefinitions ctx >=> runDeclare ctx