+++ test/Rendering/TestSchemaRendering.hs
@@ -7,16 +7,14 @@ module Rendering.TestSchemaRendering
 where
 
 import qualified Data.ByteString.Lazy.Char8 as T
-import Data.Morpheus.Document (toGraphQLDocument)
+import Data.Morpheus.Server (printSchema)
 import Relude hiding (ByteString, readFile)
 import Rendering.Schema (proxy)
 import Test.Morpheus (file, mkUrl, renderingAssertion)
 import Test.Tasty (TestTree)
 
 xyz :: b -> IO (Either String T.ByteString)
-xyz =
-  const $ pure $ Right $
-    toGraphQLDocument proxy
+xyz _ = pure $ Right $ printSchema proxy
 
 testSchemaRendering :: TestTree
 testSchemaRendering = renderingAssertion xyz (file (mkUrl "Rendering") "schema")